// ClientApp/src/app/book-management/book-management-tab/models/PublishersInformation.ts
export interface PublishersInformation {
    publishersInformationId: number;
    baseSchoolNameId: number;
    bookInformationId: number;
    publishersName: string;
    publishersAddress: string;
    publisherDate: string;
    publisherPlace: string;
    //menuPosition: string;
    isActive: boolean;
}